+++ package.json
@@ -48,6 +48,7 @@
     "body-parser": "^1.15.2",
     "bower": "^1.7.9",
     "express": "^4.14.0",
+    "express-http-proxy": "^0.10.0",
     "itemsapi": "^1.0.34",
     "itemsapi-node": "^1.0.14",
     "lodash": "^4.13.1",
